tests: Fix mypy errors in test_stream_readers_ci.py
tests\component\test_stream_readers_ci.py:180: error: Argument 1 to "_validate_frequency_data" has incompatible type "ndarray[Any, dtype[floating[_64Bit]]]"; expected "ndarray[Any, dtype[unsignedinteger[_32Bit]]]"  [arg-type]

+++ tests/component/test_stream_readers_ci.py
@@ -30,7 +30,7 @@ def _validate_count_edges_data(data: numpy.typing.NDArray[numpy.uint32]) -> None
         assert datum >= last
 
 
-def _validate_frequency_data(data: numpy.typing.NDArray[numpy.uint32]) -> None:
+def _validate_frequency_data(data: numpy.typing.NDArray[numpy.float64]) -> None:
     assert data == pytest.approx(EXPECTED_FREQUENCY, abs=EXPECTED_FREQUENCY_TOLERANCE)
